change lookup strategy for commits, taking into account merged branches and all found tags
instead of using a kind of 'git log' lookup strategy, navigate to the reachable graph of commits from HEAD
and stop at each commit found containing a version tag. If a found commit is reachable via several paths,
only the shortest path is retained. Then from the list of found tags, the one with the highest version is kept and used.

fixes #53